Commit initial ldap sni proxy server
This is the first version of a tcp reverse proxy server
for ldap connections supporting Server Name Indication (SNI)
for tls based connections.

Existing solutions with TCP TLS SNI support do not work
because they expect a direct TLS handshake but for ldap
prior to the TLS handshake the ldap extended request
LDAP_START_TLS_OID is exchanged.

This server intercepts the ldap extended request for
LDAP_START_TLS_OID pretending the server supports TLS
and detects the SNI server name from the client hello message.
The TLS connection is created using the server name.

TLS termination happens on the backend. It is not necessary
to configure any certificates in this proxy.

TODO: Add a mapping for detected SNI server name to backend(s)